Renamed MasterInterface to States.h
